Add files via upload
Business Objective: 

The main objective of this case study is to conduct an in-depth analysis of the loan dataset to identify patterns and uncover factors that indicate whether a client is likely to experience difficulties in repaying their installments. By examining the client's profile and loan attributes, the study aims to understand the driving factors behind loan default, providing valuable insights for the company's risk analytics in the banking and financial services sector.
The analysis will enable the company to take proactive measures, such as denying loans, reducing loan amounts, or offering loans to risky applicants at higher interest rates, in order to minimize financial losses and optimize the loan portfolio. By utilizing Exploratory Data Analysis (EDA) techniques, the company can gain a comprehensive understanding of the types of variables and their significance in predicting loan default.
Furthermore, the study seeks to evaluate the potential of logistic regression as a predictive model to forecast loan approval or refusal for individual clients. By training the logistic regression model on historical loan data, the company can estimate the probability of payment difficulties for new loan applicants. This predictive capability will empower the company to make informed decisions about loan applications, ensuring that deserving clients capable of repaying the loan are not rejected while mitigating the risk of default.
Overall, the objectives of this case study encompass enhancing risk assessment methodologies, improving loan approval processes, and leveraging data-driven insights to make accurate lending decisions that align with the company's financial objectives and minimize potential losses.